Move types.d.ts into index.d.ts
